move const stuff to requisitions module
