+++ corehq/apps/commtrack/const.py
@@ -61,22 +61,6 @@ ORDERED_REQUISITION_ACTIONS = (
     RequisitionActions.RECEIPTS,
 )
 
-class UserRequisitionRoles(object):
-    REQUESTER = 'commtrack_requester'
-    APPROVER = 'commtrack_approver'
-    SUPPLIER = 'commtrack_supplier'
-    RECEIVER = 'commtrack_receiver'
-
-    @classmethod
-    def get_user_role(cls, action_type):
-        return {
-            RequisitionActions.REQUEST: cls.REQUESTER,
-            RequisitionActions.APPROVAL: cls.APPROVER,
-            RequisitionActions.FULFILL: cls.SUPPLIER,
-            RequisitionActions.RECEIPTS: cls.RECEIVER,
-        }[action_type]
-
-
 class RequisitionStatus(object):
     """a const for our requisition status choices"""
     REQUESTED = "requested"
@@ -113,11 +97,3 @@ def get_commtrack_user_id(domain):
 
 USER_LOCATION_OWNER_MAP_TYPE = 'user-owner-mapping-case'
 
-def notification_template(action):
-    # this had to be a method to do translations
-    from django.utils.translation import ugettext as _
-    return {
-        RequisitionActions.APPROVAL: _('{name} has requested the following supplies: {summary}. please respond "{keyword} {loc}" to approve.'),
-        RequisitionActions.FULFILL: _('{name} should be supplied with the following supplies: {summary}. please respond "{keyword} {loc}" to confirm the order.'),
-        RequisitionActions.RECEIPTS: _('your order of {summary} is ready to be picked up. please respond with a "{keyword}" message to report receipts.'),
-    }[action]
